Allow for per-host autoplaying of flash files.
Perferences are stored in an ini style file under $XDG_CONFIG_HOME as
swfdec-mozilla.conf. The 'global' group stores default options, which
are overridden per-host by a group with the name of that host. For now
there is only the 'autoplay' boolean option.

Playing/stopping a flash file via the ui will alter the config value
for that file's host.