draw to the visual from the NPWindow
Currently the swfdec plugin draws with the visual from the browser toplevel
window which is not necessarily appropriate for the Drawable of windowless
plugins.  This can result in "BadMatch (invalid parameter attributes)" when the
depths differ or colour confusion if the depths are the same but color masks
differ.

Once this Mozilla bug is fixed, its test case can serve as a test case for this
bug:

https://bugzilla.mozilla.org/show_bug.cgi?id=445250

Even without non-unit opacity, it's possible there may be some
Xserver/application configurations where the visual on the toplevel window
differs from that used within Mozilla, which is gdk_rgb_get_colormap(), or
other browser.

Fixes bug #16717