Implement different autoplay policies, add GUI for changing between them
The three possible configuration are: always autoplay, never autoplay and check
last choice for the same host.

The idea implemented is to have the global autoplay key be not overridden by
the per host configuration. If the global autoplay key is not present the
behaviour of the actual implementation is preserved, and its default is still
to do not play automatically.